chore: added profile permission check when updating and deleting students records
- created `profile_has_student_record_update_permissions?/2` and `profile_has_student_record_delete_permissions?/2` functions in `StudentsRecords` context
- added support to `check_profile_permissions` opt in `StudentsRecords`'s `update_student_record/3` and `delete_student_record/2`
- implemented newe functions/opts in `StudentRecordOverlayComponent`
- minor adjustment in `MenuComponent` to render staff member role on profile change

+++ lib/lanttern/students_records.ex
@@ -4,15 +4,42 @@ defmodule Lanttern.StudentsRecords do
 
   # About profile permissions
 
-  The functions `list_students_records/1` and `get_student_record/2` accept the `:check_profile_permissions` option.
+  Using the `:check_profile_permissions` option, we can validate read,
+  update, and delete students records operations based on given `%Profile{}`
+  (usually from `socket.assigns.current_user.current_profile`).
 
-  A profile has permission to view a student record if it has type `"staff"`
+  ## Read permissions
+
+  Supported by `list_students_records/1` and `get_student_record/2`.
+
+  A profile has permission to read a student record if it has type `"staff"`
   and the staff member belongs to the same school as the student record, and:
 
   1. the profile has the `students_records_full_access` permission
   2. or the student record is `shared_with_school`
   3. or the staff member created the student record
   4. or the staff member is an assignee of the student record
+
+  ## Update permissions
+
+  Supported by `update_student_record/3`.
+
+  A profile has permission to update a student record if it has type `"staff"`
+  and the staff member belongs to the same school as the student record, and:
+
+  1. the profile has the `students_records_full_access` permission
+  2. or the staff member created the student record
+  3. or the staff member is an assignee of the student record
+
+  ## Delete permissions
+
+  Supported by `delete_student_record/2`.
+
+  A profile has permission to delete a student record if it has type `"staff"`
+  and the staff member belongs to the same school as the student record, and:
+
+  1. the profile has the `students_records_full_access` permission
+  2. or the staff member created the student record
 
   """
 
@@ -147,7 +174,7 @@ defmodule Lanttern.StudentsRecords do
          | opts
        ]) do
     queryable
-    |> apply_check_profile_permissions(
+    |> apply_check_profile_read_permissions(
       profile,
       "students_records_full_access" in permissions
     )
@@ -192,7 +219,7 @@ defmodule Lanttern.StudentsRecords do
   defp apply_list_students_records_opts(queryable, [_ | opts]),
     do: apply_list_students_records_opts(queryable, opts)
 
-  defp apply_check_profile_permissions(
+  defp apply_check_profile_read_permissions(
          queryable,
          %Profile{staff_member: %StaffMember{school_id: school_id}},
          true
@@ -203,7 +230,7 @@ defmodule Lanttern.StudentsRecords do
     )
   end
 
-  defp apply_check_profile_permissions(
+  defp apply_check_profile_read_permissions(
          queryable,
          %Profile{staff_member: %StaffMember{id: staff_member_id, school_id: school_id}},
          _has_full_acceess = false
@@ -223,7 +250,7 @@ defmodule Lanttern.StudentsRecords do
     )
   end
 
-  defp apply_check_profile_permissions(
+  defp apply_check_profile_read_permissions(
          queryable,
          _profile,
          _has_full_access
@@ -295,7 +322,7 @@ defmodule Lanttern.StudentsRecords do
          | opts
        ]) do
     queryable
-    |> apply_check_profile_permissions(
+    |> apply_check_profile_read_permissions(
       profile,
       "students_records_full_access" in permissions
     )
@@ -354,6 +381,7 @@ defmodule Lanttern.StudentsRecords do
   ## Options:
 
   - `:log_profile_id` - logs the operation, linked to given profile
+  - `:check_profile_permissions` - check if the user has permission to update the record
 
   ## Examples
 
@@ -365,11 +393,56 @@ defmodule Lanttern.StudentsRecords do
 
   """
   def update_student_record(%StudentRecord{} = student_record, attrs, opts \\ []) do
-    student_record
-    |> StudentRecord.changeset(attrs)
-    |> Repo.update()
-    |> StudentsRecordsLog.maybe_create_student_record_log("UPDATE", opts)
+    # for profile permissions check we get the student record from DB
+    # to prevent checking against a stale record
+    check_profile_update_permissions(
+      get_student_record(student_record.id, preloads: :assignees),
+      Keyword.get(opts, :check_profile_permissions)
+    )
+    |> case do
+      :ok ->
+        student_record
+        |> StudentRecord.changeset(attrs)
+        |> Repo.update()
+        |> StudentsRecordsLog.maybe_create_student_record_log("UPDATE", opts)
+
+      _ ->
+        {:error, %Ecto.Changeset{}}
+    end
   end
+
+  defp check_profile_update_permissions(_student_record, nil), do: :ok
+
+  defp check_profile_update_permissions(
+         %StudentRecord{} = student_record,
+         %Profile{} = profile
+       ) do
+    if profile_has_student_record_update_permissions?(student_record, profile),
+      do: :ok,
+      else: :error
+  end
+
+  defp check_profile_update_permissions(_student_record, _opt), do: :error
+
+  @doc """
+  Check if profile has permission to update student record.
+
+  Expects profile permissions, profile staff member, and student record assignees preloads.
+
+  """
+  @spec profile_has_student_record_update_permissions?(StudentRecord.t(), Profile.t()) ::
+          boolean()
+  def profile_has_student_record_update_permissions?(
+        %StudentRecord{school_id: student_record_school_id} = student_record,
+        %Profile{staff_member: %StaffMember{school_id: profile_school_id}} = profile
+      )
+      when student_record_school_id == profile_school_id do
+    "students_records_full_access" in profile.permissions or
+      student_record.created_by_staff_member_id == profile.staff_member.id or
+      Enum.any?(student_record.assignees, &(&1.id == profile.staff_member.id))
+  end
+
+  def profile_has_student_record_update_permissions?(_student_record, _profile), do: false
 
   @doc """
   Deletes a student_record.
@@ -377,6 +450,7 @@ defmodule Lanttern.StudentsRecords do
   ## Options:
 
   - `:log_profile_id` - logs the operation, linked to given profile
+  - `:check_profile_permissions` - check if the user has permission to delete the record
 
   ## Examples
 
@@ -388,9 +462,50 @@ defmodule Lanttern.StudentsRecords do
 
   """
   def delete_student_record(%StudentRecord{} = student_record, opts \\ []) do
-    Repo.delete(student_record)
-    |> StudentsRecordsLog.maybe_create_student_record_log("DELETE", opts)
+    # for profile permissions check we get the student record from DB
+    # to prevent checking against a stale record
+    check_profile_delete_permissions(
+      get_student_record(student_record.id),
+      Keyword.get(opts, :check_profile_permissions)
+    )
+    |> case do
+      :ok ->
+        Repo.delete(student_record)
+        |> StudentsRecordsLog.maybe_create_student_record_log("DELETE", opts)
+
+      _ ->
+        {:error, %Ecto.Changeset{}}
+    end
   end
+
+  defp check_profile_delete_permissions(_student_record, nil), do: :ok
+
+  defp check_profile_delete_permissions(%StudentRecord{} = student_record, %Profile{} = profile) do
+    if profile_has_student_record_delete_permissions?(student_record, profile),
+      do: :ok,
+      else: :error
+  end
+
+  defp check_profile_delete_permissions(_student_record, _opt), do: :error
+
+  @doc """
+  Check if profile has permission to delete student record.
+
+  Expects profile permissions and staff member preloads.
+
+  """
+  @spec profile_has_student_record_delete_permissions?(StudentRecord.t(), Profile.t()) ::
+          boolean()
+  def profile_has_student_record_delete_permissions?(
+        %StudentRecord{school_id: student_record_school_id} = student_record,
+        %Profile{staff_member: %StaffMember{school_id: profile_school_id}} = profile
+      )
+      when student_record_school_id == profile_school_id do
+    "students_records_full_access" in profile.permissions or
+      student_record.created_by_staff_member_id == profile.staff_member.id
+  end
+
+  def profile_has_student_record_delete_permissions?(_student_record, _profile), do: false
 
   @doc """
   Returns an `%Ecto.Changeset{}` for tracking student_record changes.
